requirements.txt
streamlit
PyMuPDF
google-generativeai
google-adk
qdrant-client
python-dotenv
pdf2image
Pillow